Refactor network service backend layout
