+++ crates/hydebar-core/src/services/network/backend/iwd/adapter.rs
