+++ crates/hydebar-core/src/services/network/backend/iwd/service_manager.rs
