+++ crates/hydebar-core/src/services/network/backend/iwd/station_diagnostic.rs
